backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
greenlet==3.2.1
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pvlib==0.12.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scipy==1.15.2
six==1.17.0
SQLAlchemy==2.0.40
python-dotenv
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
PyJWT==2.8.0
email-validator==2.1.0
Flask
Flask-SQLAlchemy
flask-socketio
eventlet
Flask-Cors
Flask-Migrate
Flask-Mail==0.9.1
msal==1.31.0
requests==2.32.3
psycopg2-binary
gunicorn
pandas
numpy
pvlib
